backend_v2/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication & Security
PyJWT==2.8.0
flask-jwt-extended==4.5.3
bcrypt==4.1.1
pydantic==2.5.0
python-dotenv==1.0.0

# Async Processing
celery==5.3.4
redis==5.0.1
flower==2.0.1  # Celery monitoring

# OCR & Image Processing
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3
ImageHash==4.3.1

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Text Processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Date/Time
python-dateutil==2.8.2

# Rate Limiting
Flask-Limiter==3.5.0

# API Documentation
flask-openapi3==3.0.0
apispec==6.3.1

# Monitoring & Logging
prometheus-flask-exporter==0.22.4
sentry-sdk[flask]==1.38.0

# Validation
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-flask==1.3.0
faker==20.1.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5

# Cloud OCR (Optional)
google-cloud-vision==3.4.5
azure-ai-formrecognizer==3.3.0

# Email
imapclient==2.3.1

# Development
ipython==8.17.2
ipdb==0.13.13
